requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.4
distro==1.9.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
multidict==6.7.1
openai==2.26.0
pillow==12.1.1
propcache==0.4.1
pyaes==1.6.1
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
pyTelegramBotAPI==4.31.0
python-dotenv==1.2.2
python-telegram-bot==22.6
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
telebot==0.0.5
Telethon==1.42.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.23.0
